Refactor inbox functions to improve code organization
