+++ bots/stress/index.ts
@@ -1,5 +1,5 @@
 import { logAndSend } from "@helpers/client";
-import { getInboxIds } from "@inboxes/gen";
+import { getInboxIds } from "@inboxes/utils";
 import {
   type Client,
   type Conversation,
