+++ inboxes/gen.ts
@@ -7,71 +7,6 @@ import {
   loadEnv,
 } from "@helpers/client";
 import { Client, type XmtpEnv } from "@xmtp/node-sdk";
-import newInboxes2 from "./2.json";
-import newInboxes5 from "./5.json";
-import newInboxes10 from "./10.json";
-import newInboxes15 from "./15.json";
-import newInboxes20 from "./20.json";
-import newInboxes25 from "./25.json";
-import newInboxes30 from "./30.json";
-
-const typedInboxes2 = newInboxes2 as InboxData[];
-const typedInboxes5 = newInboxes5 as InboxData[];
-const typedInboxes10 = newInboxes10 as InboxData[];
-const typedInboxes20 = newInboxes20 as InboxData[];
-const typedInboxes25 = newInboxes25 as InboxData[];
-const typedInboxes15 = newInboxes15 as InboxData[];
-const typedInboxes30 = newInboxes30 as InboxData[];
-
-export function getInboxByInstallationCount(
-  installationCount: number,
-  index?: number,
-) {
-  if (installationCount === 2) {
-    return index !== undefined ? typedInboxes2.slice(0, index) : typedInboxes2;
-  } else if (installationCount === 5) {
-    return index !== undefined ? typedInboxes5.slice(0, index) : typedInboxes5;
-  } else if (installationCount === 10) {
-    return index !== undefined
-      ? typedInboxes10.slice(0, index)
-      : typedInboxes10;
-  } else if (installationCount === 15) {
-    return index !== undefined
-      ? typedInboxes15.slice(0, index)
-      : typedInboxes15;
-  } else if (installationCount === 20) {
-    return index !== undefined
-      ? typedInboxes20.slice(0, index)
-      : typedInboxes20;
-  } else if (installationCount === 25) {
-    return index !== undefined
-      ? typedInboxes25.slice(0, index)
-      : typedInboxes25;
-  } else if (installationCount === 30) {
-    return index !== undefined
-      ? typedInboxes30.slice(0, index)
-      : typedInboxes30;
-  }
-  return typedInboxes2;
-}
-
-export function getRandomInboxIds(count: number) {
-  return getInboxByInstallationCount(2)
-    .sort(() => Math.random() - 0.5)
-    .slice(0, count)
-    .map((inbox) => inbox.inboxId);
-}
-
-export function getInboxIds(count: number) {
-  return getInboxByInstallationCount(2)
-    .slice(0, count)
-    .map((inbox) => inbox.inboxId);
-}
-export function getAddresses(count: number) {
-  return getInboxByInstallationCount(2)
-    .slice(0, count)
-    .map((inbox) => inbox.accountAddress);
-}
 
 // Simple progress indicator
 function showProgress(current: number, total: number, filename: string) {
@@ -102,8 +37,8 @@ function countInboxIdDuplicates(inboxes: InboxData[]): {
   };
 }
 
-// Main function to analyze all JSON files (only inboxId duplicates, and remove them)
-function analyzeAllFiles(): void {
+// Export the analysis function separately instead of running it automatically
+export function analyzeInboxFiles(): void {
   console.log(`🔍 XMTP Duplicate InboxId Counter & Remover\n`);
   console.log(`📁 Analyzing JSON files in ${INBOXES_DIR}\n`);
 
@@ -435,7 +370,7 @@ async function smartUpdate(opts: {
   if (!fs.existsSync(LOGPATH)) {
     fs.mkdirSync(LOGPATH, { recursive: true });
   }
-  analyzeAllFiles();
+  analyzeInboxFiles();
 
   let totalCreated = 0;
   let totalFailed = 0;
