+++ suites/browser/browser.test.ts
@@ -6,7 +6,7 @@ import {
   getInboxByInstallationCount,
   getInboxIds,
   getRandomInboxIds,
-} from "@inboxes/gen";
+} from "@inboxes/utils";
 import { typeOfResponse, typeofStream } from "@workers/main";
 import { getWorkers, type Worker } from "@workers/manager";
 import { beforeAll, describe, expect, it } from "vitest";
